[refactor] shell 을 위한 commands -> shell_commands로 디렉토리명 변경
