+++ src/__init__.py
@@ -1,14 +1,14 @@
-from src.commands.help import HelpCommand
-from src.commands.read import ReadCommand
-from src.commands.write import WriteCommand
-from src.commands.full_read import FullReadCommand
-from src.commands.full_write import FullWriteCommand
-from src.commands.exit import ExitCommand
-from src.commands.erase import EraseCommand
-from src.commands.flush import FlushCommand
-from src.commands.erase_range import EraseRangeCommand
+from src.shell_commands.help import HelpCommand
+from src.shell_commands.read import ReadCommand
+from src.shell_commands.write import WriteCommand
+from src.shell_commands.full_read import FullReadCommand
+from src.shell_commands.full_write import FullWriteCommand
+from src.shell_commands.exit import ExitCommand
+from src.shell_commands.erase import EraseCommand
+from src.shell_commands.flush import FlushCommand
+from src.shell_commands.erase_range import EraseRangeCommand
 
 # 테스트 시나리오
-from src.commands.full_write_and_read_compare import FullWriteAndReadCompareCommand
-from src.commands.partial_lba_write import PartialLBAWriteCommand
-from src.commands.write_read_aging import WriteReadAgingCommand
+from src.shell_commands.full_write_and_read_compare import FullWriteAndReadCompareCommand
+from src.shell_commands.partial_lba_write import PartialLBAWriteCommand
+from src.shell_commands.write_read_aging import WriteReadAgingCommand
