+++ src/shell.py
@@ -1,4 +1,4 @@
-from src.commands.command_action import CommandAction
+from src.shell_commands.command_action import CommandAction
 from src.ssd_file_manager import SSDFileManager
 from src.ssd_driver import SSDDriver
 import sys
