+++ src/shell_commands/exit.py
@@ -1,4 +1,4 @@
-from src.commands.command_action import CommandAction, InvalidArgumentException
+from src.shell_commands.command_action import CommandAction, InvalidArgumentException
 import sys
 from src.decorators import log_call
