+++ src/shell_commands/flush.py
@@ -1,4 +1,4 @@
-from src.commands.command_action import CommandAction, InvalidArgumentException
+from src.shell_commands.command_action import CommandAction, InvalidArgumentException
 
 
 class FlushCommand(CommandAction):
