+++ src/shell_commands/full_read.py
@@ -1,4 +1,4 @@
-from src.commands.command_action import CommandAction, InvalidArgumentException
+from src.shell_commands.command_action import CommandAction, InvalidArgumentException
 from src.ssd_file_manager import SSDFileManager
