+++ src/shell_commands/full_write_and_read_compare.py
@@ -1,5 +1,5 @@
 import random
-from src.commands.command_action import CommandAction, InvalidArgumentException
+from src.shell_commands.command_action import CommandAction, InvalidArgumentException
 
 
 class FullWriteAndReadCompareCommand(CommandAction):
