+++ src/shell_commands/write_read_aging.py
@@ -1,5 +1,5 @@
 import random
-from src.commands.command_action import CommandAction, InvalidArgumentException
+from src.shell_commands.command_action import CommandAction, InvalidArgumentException
 from src.ssd_file_manager import SSDFileManager
 
 class WriteReadAgingCommand(CommandAction):
