+++ tests/test_erase.py
@@ -1,7 +1,7 @@
 import pytest
 
 from src import EraseCommand
-from src.commands.command_action import InvalidArgumentException
+from src.shell_commands.command_action import InvalidArgumentException
 from src.ssd_driver import SSDDriver
