+++ tests/test_read.py
@@ -1,7 +1,7 @@
 import pytest
 
-from src.commands.command_action import InvalidArgumentException
-from src.commands.read import ReadCommand
+from src.shell_commands.command_action import InvalidArgumentException
+from src.shell_commands.read import ReadCommand
 from src.ssd_driver import SSDDriver
