+++ tests/test_runner.py
@@ -3,7 +3,7 @@ import tempfile
 from pathlib import Path
 
 from src.shell import runner_mode
-from src.commands.command_action import CommandAction
+from src.shell_commands.command_action import CommandAction
 
 
 class DummyCommand(CommandAction):
